Improved task

+++ src/main/kotlin/g3001_3100/s3001_minimum_moves_to_capture_the_queen/Solution.kt
@@ -1,6 +1,6 @@
 package g3001_3100.s3001_minimum_moves_to_capture_the_queen
 
-// #Medium #Array #Enumeration #2024_02_25_Time_128_ms_(94.59%)_Space_33.7_MB_(83.78%)
+// #Medium #Array #Enumeration #2024_02_25_Time_128_ms_(94.59%)_Space_34_MB_(48.65%)
 
 import kotlin.math.abs
 
@@ -14,8 +14,11 @@ class Solution {
                 return 2
             }
             return 1
-        } else if (abs((c - e).toDouble()) == abs((d - f).toDouble())) {
-            if (abs((a - c).toDouble()) == abs((b - d).toDouble()) && abs((e - a).toDouble()) == abs((f - b).toDouble()) && (a > e && a < c || a > c && a < e)) {
+        } else if (abs(c - e) == abs(d - f)) {
+            if (abs(a - c) == abs(b - d) &&
+                abs(e - a) == abs(f - b) &&
+                (a > e && a < c || a > c && a < e)
+            ) {
                 return 2
             }
             return 1
